requirements.txt
Django
djangorestframework
drf-spectacular
django-cors-headers
psycopg2-binary
python-decouple
django-storages
boto3
requests
google-api-python-client
Pillow
django-ckeditor
djangorestframework-simplejwt
django-filter
django-anymail
django-debug-toolbar
python-dotenv
django-celery-beat
celery
redis
drf-yasg
whitenoise
gunicorn
pytest
pytest-django
